@@ -1,27 +1,73 @@
 [tool.poetry]
-authors = ["Your Name <you@example.com>"]
+authors = ["Robert Kirberich"]
 description = ""
 name = "gerber_to_scad"
 version = "0.1.0"
 
+include = [
+    { path = "src/gerber_to_scad", format = ["sdist", "wheel"] },
+    { path = "src/gts_service", format = ["sdist", "wheel"] }
+]
+
+[tool.poetry.scripts]
+g2s = "gerber_to_scad.main:main"
+service = "gts_service.main:main"
+
 [tool.poetry.dependencies]
+python = "^3.10"
+pcb-tools = "^0.1.6"
+scipy = "^1.15.0"
+solidpython = "^1.1.3"
+
+[tool.poetry.group.dev.dependencies]
+ruff = ">0.3.0"
+mypy = ">1.14.0"
+
+[tool.poetry.group.service.dependencies]
 Django = "^4.2"
 django-bootstrap3 = "^23.1"
 django-environ = "^0.10.0"
 gunicorn = "^20.1.0"
-pcb-tools = "^0.1.6"
-python = "^3.9,<3.11"
-scipy = "^1.10.1"
-solidpython = "^1.1.3"
-taskipy = "^1.10.4"
-
-[tool.poetry.dev-dependencies]
-black = "^23.3.0"
 
 [build-system]
 build-backend = "poetry.core.masonry.api"
 requires = ["poetry-core>=1.0.0"]
 
-[tool.taskipy.tasks]
-g2s = "python gerber_to_scad/gerber_to_scad.py"
-service = "gunicorn --workers 1 --threads 8 gts_service.wsgi"
+[tool.ruff.lint.isort]
+required-imports = ["from __future__ import annotations"]
+
+[tool.mypy]
+python_version = "3.8"
+files = "src"
+
+strict = false
+
+warn_return_any = false
+warn_redundant_casts = false
+warn_unused_ignores = false
+warn_unreachable = false
+
+disallow_untyped_defs = false
+no_implicit_optional = false
+
+show_error_context = false
+show_error_codes = false
+show_error_code_links = false
+show_column_numbers = false
+pretty = false
+
+exclude = [
+    "^tests/resources/*$"
+]
+
+[[tool.mypy.overrides]]
+module = [
+    "django.*",
+    "environ",
+    "scipy.*",
+    "gerber.*",
+    "solid",
+
+]
+ignore_missing_imports = true
+
