@@ -1,14 +1,18 @@
-from django import forms
+from __future__ import annotations
 
+from django import forms
 
 forms.Widget()
 
+
 class UploadForm(forms.Form):
-    solderpaste_file = forms.FileField(label="Solder paste layer file")
+    solderpaste_file = forms.FileField(
+        label="Solder paste layer file, ends with '.gtp'"
+    )
     outline_file = forms.FileField(
-        label="Outline layer file (Optional)",
+        label="Outline layer file, ends with '.gko' (Optional)",
         required=False,
-        help_text='Outline of the board used to create the stencil. To create a stencil of different size, use the options below instead.'
+        help_text="Outline of the board used to create the stencil. To create a stencil of different size, use the options below instead.",
     )
     stencil_thickness = forms.FloatField(
         label="Thickness (in mm) of the stencil. Make sure this is a multiple of the layer height you use for printing.",
@@ -22,7 +26,7 @@ class UploadForm(forms.Form):
         label="Include a ledge around half the outline of the board, to allow aligning the stencil more easily.",
         initial=True,
         required=False,
-        widget=forms.CheckboxInput(attrs={'class': 'collapse-checkbox include-ledge'}),
+        widget=forms.CheckboxInput(attrs={"class": "collapse-checkbox include-ledge"}),
     )
     ledge_thickness = forms.FloatField(
         label="Thickness of the stencil ledge. This should be less than the thickness of the PCB.",
@@ -31,7 +35,7 @@ class UploadForm(forms.Form):
     increase_hole_size_by = forms.FloatField(
         label="Increase the size of all holes in the stencil by this amount (in mm). Use this if you find holes get printed smaller than they should.",
         initial=0.0,
-        help_text='Use negative values to increase the space between tightly spaced pads at the cost of making the pads smaller.'
+        help_text="Use negative values to increase the space between tightly spaced pads at the cost of making the pads smaller.",
     )
     simplify_regions = forms.BooleanField(
         label="Replace regions (usually rounded rectangles) with bounding boxes. <br/> Use this if you find the processing takes extremely long, but note it might have unintended effects other than removing rounded corners.",
@@ -60,10 +64,10 @@ class UploadForm(forms.Form):
     )
 
     include_frame = forms.BooleanField(
-        label='Add a thicker frame around the stencil. Allows using the stencil in consistently-sized fixtures.',
+        label="Add a thicker frame around the stencil. Allows using the stencil in consistently-sized fixtures.",
         initial=False,
         required=False,
-        widget=forms.CheckboxInput(attrs={'class': 'collapse-checkbox include-frame'}),
+        widget=forms.CheckboxInput(attrs={"class": "collapse-checkbox include-frame"}),
     )
     frame_width = forms.FloatField(
         label="Width of the stencil frame (in mm).",
