@@ -1,7 +1,9 @@
-"""gts URL Configuration
+"""
+gts URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/1.9/topics/http/urls/
+
 Examples:
 Function views
     1. Add an import:  from my_app import views
@@ -12,7 +14,11 @@ Class-based views
 Including another URLconf
     1. Import the include() function: from django.conf.urls import url, include
     2. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
+
 """
+
+from __future__ import annotations
+
 from django.urls import path
 
 from gts_service import views
